requirements.txt
fastapi
uvicorn
aiogram==2.23.1
psycopg2-binary
requests
beautifulsoup4
openai
python-dotenv
jinja2